+++ forge/src/main/java/earth/terrarium/adastra/datagen/provider/client/ModLangProvider.java
@@ -9,6 +9,7 @@ import earth.terrarium.adastra.common.registry.ModFluids;
 import earth.terrarium.adastra.common.registry.ModItems;
 import earth.terrarium.botarium.common.registry.fluid.BotariumFlowingFluid;
 import net.minecraft.data.PackOutput;
+import net.minecraft.network.chat.Component;
 import net.minecraft.world.item.BlockItem;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.level.block.Block;
@@ -55,7 +56,6 @@ public class ModLangProvider extends LanguageProvider {
                     .getPath()
                     .replace("_", " "))));
 
-
         ModFluids.FLUIDS.stream()
             .forEach(entry -> addFluid(entry,
                 StringUtils.capitaliseAllWords(entry
@@ -80,15 +80,15 @@ public class ModLangProvider extends LanguageProvider {
         add("death.attack.acid_rain", "%1$s was melted by acid rain");
         add("death.attack.acid_rain.player", "%1$s was melted by acid rain whilst trying to escape %2$s");
 
-        add(ConstantComponents.TOGGLE_SUIT_FLIGHT_KEY.getString(), "Toggle suit flight");
-        add(ConstantComponents.OPEN_RADIO_KEY.getString(), "Open radio");
-        add(ConstantComponents.AD_ASTRA_CATEGORY.getString(), "Ad Astra");
+        add(ConstantComponents.TOGGLE_SUIT_FLIGHT_KEY, "Toggle suit flight");
+        add(ConstantComponents.OPEN_RADIO_KEY, "Open radio");
+        add(ConstantComponents.AD_ASTRA_CATEGORY, "Ad Astra");
 
-        add(ConstantComponents.SUIT_FLIGHT_ENABLED.getString(), "Suit flight enabled");
-        add(ConstantComponents.SUIT_FLIGHT_DISABLED.getString(), "Suit flight disabled");
+        add(ConstantComponents.SUIT_FLIGHT_ENABLED, "Suit flight enabled");
+        add(ConstantComponents.SUIT_FLIGHT_DISABLED, "Suit flight disabled");
 
-        add(ConstantComponents.TRUE.getString(), "✔");
-        add(ConstantComponents.FALSE.getString(), "✘");
+        add(ConstantComponents.TRUE, "✔");
+        add(ConstantComponents.FALSE, "✘");
         add("text.ad_astra.temperature", "%s°C");
         add("text.ad_astra.gravity", "%sm/s²");
         add("text.ad_astra.radio.none", "No station");
@@ -123,14 +123,14 @@ public class ModLangProvider extends LanguageProvider {
         add("tooltip.ad_astra.fluid_per_tick", "%s 🪣/t");
 
         add("tooltip.ad_astra.blocks_distributed", "%s/%s");
-        add(ConstantComponents.SHOW.getString(), "Show");
-        add(ConstantComponents.HIDE.getString(), "Hide");
-        add(ConstantComponents.OXYGEN_DISTRIBUTION_AREA.getString(), "Show where the oxygen is being distributed");
+        add(ConstantComponents.SHOW, "Show");
+        add(ConstantComponents.HIDE, "Hide");
+        add(ConstantComponents.OXYGEN_DISTRIBUTION_AREA, "Show where the oxygen is being distributed");
 
-        add(ConstantComponents.CLEAR_FLUID_TANK.getString(), "Shift-right-click to clear");
+        add(ConstantComponents.CLEAR_FLUID_TANK, "Shift-right-click to clear");
 
-        add(ConstantComponents.SIDE_CONFIG.getString(), "Side Config");
-        add(ConstantComponents.REDSTONE_CONTROL.getString(), "Redstone Control");
+        add(ConstantComponents.SIDE_CONFIG, "Side Config");
+        add(ConstantComponents.REDSTONE_CONTROL, "Redstone Control");
 
         add("tooltip.ad_astra.redstone_control.always_on", "Always on");
         add("tooltip.ad_astra.redstone_control.on_when_powered", "On when powered");
@@ -138,67 +138,67 @@ public class ModLangProvider extends LanguageProvider {
         add("tooltip.ad_astra.redstone_control.never_on", "Never on");
         add("tooltip.ad_astra.redstone_control.mode", "Mode: %s");
 
-        add(ConstantComponents.ACTIVE.getString(), "Active");
-        add(ConstantComponents.INACTIVE.getString(), "Inactive");
+        add(ConstantComponents.ACTIVE, "Active");
+        add(ConstantComponents.INACTIVE, "Inactive");
 
-        add(ConstantComponents.SEQUENTIAL.getString(), "Sequential");
-        add(ConstantComponents.ROUND_ROBIN.getString(), "Round Robin");
+        add(ConstantComponents.SEQUENTIAL, "Sequential");
+        add(ConstantComponents.ROUND_ROBIN, "Round Robin");
 
-        add(ConstantComponents.CAPACITOR_ENABLED.getString(), "Capacitor enabled");
-        add(ConstantComponents.CAPACITOR_DISABLED.getString(), "Capacitor disabled");
-        add(ConstantComponents.CHANGE_MODE_SEQUENTIAL.getString(), "Set capacitor mode to \"Sequential\"");
-        add(ConstantComponents.CHANGE_MODE_ROUND_ROBIN.getString(), "Set capacitor mode to \"Round Robin\"");
+        add(ConstantComponents.CAPACITOR_ENABLED, "Capacitor enabled");
+        add(ConstantComponents.CAPACITOR_DISABLED, "Capacitor disabled");
+        add(ConstantComponents.CHANGE_MODE_SEQUENTIAL, "Set capacitor mode to \"Sequential\"");
+        add(ConstantComponents.CHANGE_MODE_ROUND_ROBIN, "Set capacitor mode to \"Round Robin\"");
 
-        add(ConstantComponents.SHIFT_DESCRIPTION.getString(), "Hold SHIFT for more information");
-        add(ConstantComponents.TI_69_INFO.getString(), "Displays important information. Right-click to change the current app.");
-        add(ConstantComponents.ETRIONIC_CAPACITOR_INFO.getString(), "Stores energy. Right-click to toggle. Shift-right-click to change the distribution mode.");
-        add(ConstantComponents.WRENCH_INFO.getString(), "Used to configure pipes and various other blocks.");
-        add(ConstantComponents.ZIP_GUN_INFO.getString(), "Propels you forward. Use a second one in your offhand to propel even further. Excellent for maneuvering in zero-gravity.");
-        add(ConstantComponents.GAS_TANK_INFO.getString(), "Stores fluids and gases. Right-click to distribute into your inventory.");
-        add(ConstantComponents.CABLE_INFO.getString(), "Transfers energy. Right-click with a wrench to change the distribution mode.");
-        add(ConstantComponents.FLUID_PIPE_INFO.getString(), "Transfers fluids. Right-click with a wrench to change the distribution mode.");
-        add(ConstantComponents.CABLE_DUCT.getString(), "Works like cables, but fully-conceals oxygen, preventing leaks.");
-        add(ConstantComponents.FLUID_DUCT_INFO.getString(), "Works like fluid pipes, but fully-conceals oxygen, preventing leaks.");
-        add(ConstantComponents.SPACE_SUIT_INFO.getString(), "Supplies the wearer with oxygen and protects them from the extreme-cold of space.");
-        add(ConstantComponents.NETHERITE_SPACE_SUIT_INFO.getString(), "Works like the Space Suit, with the added advantage of protecting the wearer from extreme heat.");
-        add(ConstantComponents.JET_SUIT_INFO.getString(), "Works like the Netherite Space Suit, with jets installed to propel the wearer forward with extreme force.");
-        add(ConstantComponents.SLIDING_DOOR_INFO.getString(), "A large 3x3 door. Shift-right-click with a wrench to lock it. Locked doors can only be opened with redstone.");
-        add(ConstantComponents.FLAG_INFO.getString(), "A flag with your face on it. Right-click to add an image from a URL.");
-        add(ConstantComponents.LAUNCH_PAD_INFO.getString(), "Launches rockets. Power the center part with redstone to automatically launch placed rockets.");
-        add(ConstantComponents.GLOBE_INFO.getString(), "Right-click to spin the globe. Power it with redstone to make it constantly spin.");
-        add(ConstantComponents.VENT_INFO.getString(), "Allows oxygen to flow through.");
+        add(ConstantComponents.SHIFT_DESCRIPTION, "Hold SHIFT for more information");
+        add(ConstantComponents.TI_69_INFO, "Displays important information. Right-click to change the current app.");
+        add(ConstantComponents.ETRIONIC_CAPACITOR_INFO, "Stores energy. Right-click to toggle. Shift-right-click to change the distribution mode.");
+        add(ConstantComponents.WRENCH_INFO, "Used to configure pipes and various other blocks.");
+        add(ConstantComponents.ZIP_GUN_INFO, "Propels you forward. Use a second one in your offhand to propel even further. Excellent for maneuvering in zero-gravity.");
+        add(ConstantComponents.GAS_TANK_INFO, "Stores fluids and gases. Right-click to distribute into your inventory.");
+        add(ConstantComponents.CABLE_INFO, "Transfers energy. Right-click with a wrench to change the distribution mode.");
+        add(ConstantComponents.FLUID_PIPE_INFO, "Transfers fluids. Right-click with a wrench to change the distribution mode.");
+        add(ConstantComponents.CABLE_DUCT, "Works like cables, but fully-conceals oxygen, preventing leaks.");
+        add(ConstantComponents.FLUID_DUCT_INFO, "Works like fluid pipes, but fully-conceals oxygen, preventing leaks.");
+        add(ConstantComponents.SPACE_SUIT_INFO, "Supplies the wearer with oxygen and protects them from the extreme-cold of space.");
+        add(ConstantComponents.NETHERITE_SPACE_SUIT_INFO, "Works like the Space Suit, with the added advantage of protecting the wearer from extreme heat.");
+        add(ConstantComponents.JET_SUIT_INFO, "Works like the Netherite Space Suit, with jets installed to propel the wearer forward with extreme force.");
+        add(ConstantComponents.SLIDING_DOOR_INFO, "A large 3x3 door. Shift-right-click with a wrench to lock it. Locked doors can only be opened with redstone.");
+        add(ConstantComponents.FLAG_INFO, "A flag with your face on it. Right-click to add an image from a URL.");
+        add(ConstantComponents.LAUNCH_PAD_INFO, "Launches rockets. Power the center part with redstone to automatically launch placed rockets.");
+        add(ConstantComponents.GLOBE_INFO, "Right-click to spin the globe. Power it with redstone to make it constantly spin.");
+        add(ConstantComponents.VENT_INFO, "Allows oxygen to flow through.");
 
-        add(ConstantComponents.ROCKET_INFO.getString(), "A rocket used for interplanetary travel. Must be placed on a launch pad. Shift-right-click to access its fuel and inventory.");
-        add(ConstantComponents.ROVER_INFO.getString(), "An all-terrain vehicle. Shift-right-click to access its inventory. Click on the screen to access the rover's radio. This radio can play real radio stations.");
+        add(ConstantComponents.ROCKET_INFO, "A rocket used for interplanetary travel. Must be placed on a launch pad. Shift-right-click to access its fuel and inventory.");
+        add(ConstantComponents.ROVER_INFO, "An all-terrain vehicle. Shift-right-click to access its inventory. Click on the screen to access the rover's radio. This radio can play real radio stations.");
 
-        add(ConstantComponents.COAL_GENERATOR_INFO.getString(), "The simplest of generators. Generates energy from burnable materials.");
-        add(ConstantComponents.COMPRESSOR_INFO.getString(), "Compresses materials into plates.");
-        add(ConstantComponents.NASA_WORKBENCH_INFO.getString(), "Used to craft rockets.");
-        add(ConstantComponents.FUEL_REFINERY_INFO.getString(), "Refines fuel from oil.");
-        add(ConstantComponents.OXYGEN_LOADER_INFO.getString(), "Converts water into oxygen. Use it to fill up space suits and gas tanks.");
-        add(ConstantComponents.SOLAR_PANEL_INFO.getString(), "Generates energy from the sun during the dat. Note that the energy generated is dependent on the planet.");
-        add(ConstantComponents.WATER_PUMP_INFO.getString(), "Pumps water. Ensure a water source is below the pump for it to function.");
-        add(ConstantComponents.OXYGEN_DISTRIBUTOR_INFO.getString(), "Distributes oxygen and regulates the temperature. Use it to create livable environments.");
-        add(ConstantComponents.ENERGIZER_INFO.getString(), "Stores a large amount of energy. Right-click with an item to charge it. Retains its charge when broken.");
-        add(ConstantComponents.CRYO_FREEZER_INFO.getString(), "Converts ice materials into cryo fuel.");
-        add(ConstantComponents.OXYGEN_SENSOR_INFO.getString(), "Emits a redstone signal when oxygen is detected. Shift-right-click to invert the signal.");
+        add(ConstantComponents.COAL_GENERATOR_INFO, "The simplest of generators. Generates energy from burnable materials.");
+        add(ConstantComponents.COMPRESSOR_INFO, "Compresses materials into plates.");
+        add(ConstantComponents.NASA_WORKBENCH_INFO, "Used to craft rockets.");
+        add(ConstantComponents.FUEL_REFINERY_INFO, "Refines fuel from oil.");
+        add(ConstantComponents.OXYGEN_LOADER_INFO, "Converts water into oxygen. Use it to fill up space suits and gas tanks.");
+        add(ConstantComponents.SOLAR_PANEL_INFO, "Generates energy from the sun during the dat. Note that the energy generated is dependent on the planet.");
+        add(ConstantComponents.WATER_PUMP_INFO, "Pumps water. Ensure a water source is below the pump for it to function.");
+        add(ConstantComponents.OXYGEN_DISTRIBUTOR_INFO, "Distributes oxygen and regulates the temperature. Use it to create livable environments.");
+        add(ConstantComponents.ENERGIZER_INFO, "Stores a large amount of energy. Right-click with an item to charge it. Retains its charge when broken.");
+        add(ConstantComponents.CRYO_FREEZER_INFO, "Converts ice materials into cryo fuel.");
+        add(ConstantComponents.OXYGEN_SENSOR_INFO, "Emits a redstone signal when oxygen is detected. Shift-right-click to invert the signal.");
 
         add("tooltip.ad_astra.energy_transfer_tick", "Transfer: %s ⚡/t");
         add("tooltip.ad_astra.fluid_transfer_tick", "Transfer: %s 🪣/t");
 
-        add(ConstantComponents.UPGRADES.getString(), "Upgrades");
+        add(ConstantComponents.UPGRADES, "Upgrades");
         add("tooltip.ad_astra.upgrades.entry", "- %sx %s");
         add("upgrade.ad_astra.speed", "Speed Upgrade");
         add("upgrade.ad_astra.capacity", "Capacity Upgrade");
 
-        add(ConstantComponents.SIDE_CONFIG_SLOTS.getString(), "Slots");
-        add(ConstantComponents.SIDE_CONFIG_ENERGY.getString(), "Energy");
-        add(ConstantComponents.SIDE_CONFIG_FLUID.getString(), "Fluid");
-        add(ConstantComponents.SIDE_CONFIG_INPUT_SLOTS.getString(), "Input slots");
-        add(ConstantComponents.SIDE_CONFIG_OUTPUT_SLOTS.getString(), "Output slots");
-        add(ConstantComponents.SIDE_CONFIG_EXTRACTION_SLOTS.getString(), "Extraction slots");
-        add(ConstantComponents.SIDE_CONFIG_INPUT_FLUID.getString(), "Input fluid");
-        add(ConstantComponents.SIDE_CONFIG_OUTPUT_FLUID.getString(), "Output fluid");
+        add(ConstantComponents.SIDE_CONFIG_SLOTS, "Slots");
+        add(ConstantComponents.SIDE_CONFIG_ENERGY, "Energy");
+        add(ConstantComponents.SIDE_CONFIG_FLUID, "Fluid");
+        add(ConstantComponents.SIDE_CONFIG_INPUT_SLOTS, "Input slots");
+        add(ConstantComponents.SIDE_CONFIG_OUTPUT_SLOTS, "Output slots");
+        add(ConstantComponents.SIDE_CONFIG_EXTRACTION_SLOTS, "Extraction slots");
+        add(ConstantComponents.SIDE_CONFIG_INPUT_FLUID, "Input fluid");
+        add(ConstantComponents.SIDE_CONFIG_OUTPUT_FLUID, "Output fluid");
 
         add("side_config.ad_astra.title", "Side Config [%s]");
         add("side_config.ad_astra.type.type", "Type: %s");
@@ -228,35 +228,35 @@ public class ModLangProvider extends LanguageProvider {
         add("direction.ad_astra.relative.south", "Back");
         add("direction.ad_astra.relative.west", "Right");
 
-        add(ConstantComponents.NEXT.getString(), "Next");
-        add(ConstantComponents.PREVIOUS.getString(), "Previous");
-        add(ConstantComponents.RESET_TO_DEFAULT.getString(), "Reset to default");
+        add(ConstantComponents.NEXT, "Next");
+        add(ConstantComponents.PREVIOUS, "Previous");
+        add(ConstantComponents.RESET_TO_DEFAULT, "Reset to default");
 
-        add(ConstantComponents.DETECTOR_INVERTED_TRUE.getString(), "Set to inverted");
-        add(ConstantComponents.DETECTOR_INVERTED_FALSE.getString(), "Set to normal");
-        add(ConstantComponents.DETECTOR_OXYGEN_MODE.getString(), "Set to oxygen detection mode");
-        add(ConstantComponents.DETECTOR_GRAVITY_MODE.getString(), "Set to normal gravity detection mode");
-        add(ConstantComponents.DETECTOR_TEMPERATURE_MODE.getString(), "Set to safe temperature detection Mode");
+        add(ConstantComponents.DETECTOR_INVERTED_TRUE, "Set to inverted");
+        add(ConstantComponents.DETECTOR_INVERTED_FALSE, "Set to normal");
+        add(ConstantComponents.DETECTOR_OXYGEN_MODE, "Set to oxygen detection mode");
+        add(ConstantComponents.DETECTOR_GRAVITY_MODE, "Set to normal gravity detection mode");
+        add(ConstantComponents.DETECTOR_TEMPERATURE_MODE, "Set to safe temperature detection Mode");
 
-        add(ConstantComponents.FLAG_URL.getString(), "Flag Url (https://i.imgur.com/urURL.png)");
-        add(ConstantComponents.CONFIRM.getString(), "Confirm");
-        add(ConstantComponents.NOT_THE_OWNER.getString(), "You do not own this flag!");
+        add(ConstantComponents.FLAG_URL, "Flag Url (https://i.imgur.com/urURL.png)");
+        add(ConstantComponents.CONFIRM, "Confirm");
+        add(ConstantComponents.NOT_THE_OWNER, "You do not own this flag!");
 
-        add(ConstantComponents.NOT_ENOUGH_FUEL.getString(), "Not enough fuel!");
-        add(ConstantComponents.INVALID_LAUNCHING_DIMENSION.getString(), "You cannot launch from this dimension!");
+        add(ConstantComponents.NOT_ENOUGH_FUEL, "Not enough fuel!");
+        add(ConstantComponents.INVALID_LAUNCHING_DIMENSION, "You cannot launch from this dimension!");
 
-        add(ConstantComponents.CATALOG.getString(), "Catalog");
-        add(ConstantComponents.LAND.getString(), "Land");
-        add(ConstantComponents.SPACE_STATION.getString(), "Space Station");
-        add(ConstantComponents.CONSTRUCT.getString(), "Construct");
+        add(ConstantComponents.CATALOG, "Catalog");
+        add(ConstantComponents.LAND, "Land");
+        add(ConstantComponents.SPACE_STATION, "Space Station");
+        add(ConstantComponents.CONSTRUCT, "Construct");
         add("tooltip.ad_astra.land", "Land on %s at [%s, %s]");
         add("tooltip.ad_astra.space_station_land", "Land on space station in %s at [%s, %s]");
         add("tooltip.ad_astra.construct_space_station_at", "Construct space station in %s at [%s, %s]");
 
-        add(ConstantComponents.CONSTRUCTION_COST.getString(), "Construction Cost:");
+        add(ConstantComponents.CONSTRUCTION_COST, "Construction Cost:");
         add("tooltip.ad_astra.requirement", "%s/%s %s");
 
-        add(ConstantComponents.SPACE_STATION_ALREADY_EXISTS.getString(), "A space station already exists at this location!");
+        add(ConstantComponents.SPACE_STATION_ALREADY_EXISTS, "A space station already exists at this location!");
 
         add("solar_system.ad_astra.solar_system", "Solar System");
         add("solar_system.ad_astra.proxima_centauri", "Proxima Centauri");
@@ -311,6 +311,9 @@ public class ModLangProvider extends LanguageProvider {
         add("tag.item.ad_astra.raw_calorite", "Raw Calorite");
         add("tag.item.ad_astra.calorite_blocks", "Calorite Blocks");
         add("tag.item.ad_astra.raw_calorite_blocks", "Raw Calorite Blocks");
+        add("tag.item.ad_astra.glacian_logs", "Glacian Logs");
+        add("tag.item.ad_astra.aeronos_caps", "Aeronos Caps");
+        add("tag.item.ad_astra.strophar_caps", "Strophar Caps");
 
         add("config.ad_astra.allowFlagImages", "Allow flag images");
         add("config.ad_astra.launchFromAnywhere", "Launch from anywhere");
@@ -322,6 +325,10 @@ public class ModLangProvider extends LanguageProvider {
         add("subtitles.ad_astra.block.sliding_door_close", "Sliding door closes");
         add("subtitles.ad_astra.entity.oxygen_intake", "Oxygen intake");
         add("subtitles.ad_astra.entity.oxygen_outtake", "Oxygen outtake");
+    }
+
+    public void add(Component component, String name) {
+        add(component.getString(), name);
     }
 
     public void addFluid(Supplier<? extends Fluid> key, String name) {